backend/pyproject.toml
[tool.poetry]
name = "lightbikeshop"
version = "0.0.0"
description = ""
authors = ["pharci <lesha.belikov2005@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic-settings = "^2.7.1"
aiogram = "^3.16.0"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
itsdangerous = "^2.2.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
pillow = "^11.1.0"
setuptools = "^75.8.0"
sqlalchemy-file = "^0.6.0"
jinja2 = "^3.1.5"
fasteners = "^0.19"
python-multipart = "^0.0.20"
fastadmin = {extras = ["fastapi", "sqlalchemy"], version = "^0.2.19"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
